chore: refactor handlers and routes
