+++ server/handlers/deviceConfigurationHandler/index.js
@@ -1,37 +1,14 @@
-const Joi = require('joi');
 const {
   getFilterAndLimitOptions,
   generatePrevAndNext,
-} = require('../utils/helper');
-const DeviceConfigurationService = require('../services/DeviceConfigurationService');
-const HttpError = require('../utils/HttpError');
-
-const deviceConfigurationPostSchema = Joi.object({
-  id: Joi.string().uuid().required(),
-  device_identifier: Joi.string().required(),
-  brand: Joi.string().required(),
-  model: Joi.string().required(),
-  device: Joi.string().required(),
-  serial: Joi.string().allow('', null),
-  hardware: Joi.string().required(),
-  manufacturer: Joi.string().required(),
-  app_build: Joi.string().required(),
-  app_version: Joi.string().required(),
-  os_version: Joi.string().required(),
-  sdk_version: Joi.string().required(),
-  logged_at: Joi.string().isoDate().required(),
-  bulk_pack_file_name: Joi.string(),
-}).unknown(false);
-
-const deviceConfigurationIdParamSchema = Joi.object({
-  device_configuration_id: Joi.string().uuid().required(),
-}).unknown(false);
-
-const deviceConfigurationGetQuerySchema = Joi.object({
-  offset: Joi.number().integer().greater(-1),
-  limit: Joi.number().integer().greater(0),
-  bulk_pack_file_name: Joi.string(),
-});
+} = require('../../utils/helper');
+const DeviceConfigurationService = require('../../services/DeviceConfigurationService');
+const HttpError = require('../../utils/HttpError');
+const {
+  deviceConfigurationGetQuerySchema,
+  deviceConfigurationIdParamSchema,
+  deviceConfigurationPostSchema,
+} = require('./schemas');
 
 const deviceConfigurationPost = async function (req, res) {
   await deviceConfigurationPostSchema.validateAsync(req.body, {
@@ -40,10 +17,8 @@ const deviceConfigurationPost = async function (req, res) {
 
   const deviceConfigurationService = new DeviceConfigurationService();
 
-  const {
-    deviceConfiguration,
-    status,
-  } = await deviceConfigurationService.createDeviceConfiguration(req.body);
+  const { deviceConfiguration, status } =
+    await deviceConfigurationService.createDeviceConfiguration(req.body);
 
   res.status(status).json(deviceConfiguration);
 };
@@ -56,10 +31,11 @@ const deviceConfigurationGet = async function (req, res) {
   const { filter, limitOptions } = getFilterAndLimitOptions(req.query);
   const deviceConfigurationService = new DeviceConfigurationService();
 
-  const deviceConfigurations = await deviceConfigurationService.getDeviceConfigurations(
-    filter,
-    limitOptions,
-  );
+  const deviceConfigurations =
+    await deviceConfigurationService.getDeviceConfigurations(
+      filter,
+      limitOptions,
+    );
   const count = await deviceConfigurationService.getDeviceConfigurationsCount(
     filter,
   );
@@ -86,9 +62,10 @@ const deviceConfigurationSingleGet = async function (req, res) {
   });
 
   const deviceConfigurationService = new DeviceConfigurationService();
-  const deviceConfiguration = await deviceConfigurationService.getDeviceConfigurationById(
-    req.params.device_configuration_id,
-  );
+  const deviceConfiguration =
+    await deviceConfigurationService.getDeviceConfigurationById(
+      req.params.device_configuration_id,
+    );
 
   if (!deviceConfiguration?.id) {
     throw new HttpError(
