+++ server/handlers/replayEventHandler/schemas.js
@@ -1,0 +1,7 @@
+const Joi = require('joi');
+
+const replayEventAPISchema = Joi.object({
+  status: Joi.string().valid('raised', 'received').required(),
+});
+
+module.exports = { replayEventAPISchema };
