+++ server/handlers/sessionHandler/index.js
@@ -1,31 +1,14 @@
-const Joi = require('joi');
-const SessionService = require('../services/SessionService');
-const HttpError = require('../utils/HttpError');
+const SessionService = require('../../services/SessionService');
+const HttpError = require('../../utils/HttpError');
 const {
   getFilterAndLimitOptions,
   generatePrevAndNext,
-} = require('../utils/helper');
-
-const sessionGetQuerySchema = Joi.object({
-  offset: Joi.number().integer().greater(-1),
-  limit: Joi.number().integer().greater(0),
-  bulk_pack_file_name: Joi.string(),
-});
-
-const sessionPostSchema = Joi.object({
-  id: Joi.string().uuid().required(),
-  device_configuration_id: Joi.string().uuid().required(),
-  originating_wallet_registration_id: Joi.string().uuid().required(),
-  target_wallet: Joi.string(),
-  check_in_photo_url: Joi.string().uri(),
-  track_url: Joi.string().uri(),
-  organization: Joi.string().allow(null, ''),
-  bulk_pack_file_name: Joi.string(),
-}).unknown(false);
-
-const sessionIdParamSchema = Joi.object({
-  session_id: Joi.string().uuid().required(),
-}).unknown(false);
+} = require('../../utils/helper');
+const {
+  sessionGetQuerySchema,
+  sessionPostSchema,
+  sessionIdParamSchema,
+} = require('./schemas');
 
 const sessionPost = async function (req, res) {
   await sessionPostSchema.validateAsync(req.body, {
