+++ server/handlers/walletRegistrationHandler/index.js
@@ -1,36 +1,14 @@
-const Joi = require('joi');
-const HttpError = require('../utils/HttpError');
-const WalletRegistrationService = require('../services/WalletRegistrationService');
+const HttpError = require('../../utils/HttpError');
+const WalletRegistrationService = require('../../services/WalletRegistrationService');
 const {
   getFilterAndLimitOptions,
   generatePrevAndNext,
-} = require('../utils/helper');
-
-const walletRegistrationPostSchema = Joi.object({
-  id: Joi.string().uuid().required(),
-  wallet: Joi.string().required(),
-  user_photo_url: Joi.string().uri().required(),
-  grower_account_id: Joi.string().uuid().required(),
-  first_name: Joi.string().required(),
-  last_name: Joi.string().required(),
-  phone: Joi.string().allow(null),
-  email: Joi.string().pattern(/.*@.*/).allow(null),
-  lat: Joi.number().required().min(-90).max(90).required(),
-  lon: Joi.number().required().min(-180).max(180).required(),
-  registered_at: Joi.string().isoDate(),
-  v1_legacy_organization: Joi.string().allow(null, ''),
-  bulk_pack_file_name: Joi.string(),
-}).unknown(false);
-
-const walletRegistrationIdParamSchema = Joi.object({
-  wallet_registration_id: Joi.string().uuid().required(),
-}).unknown(false);
-
-const walletRegistrationGetQuerySchema = Joi.object({
-  offset: Joi.number().integer().greater(-1),
-  limit: Joi.number().integer().greater(0),
-  bulk_pack_file_name: Joi.string(),
-});
+} = require('../../utils/helper');
+const {
+  walletRegistrationPostSchema,
+  walletRegistrationIdParamSchema,
+  walletRegistrationGetQuerySchema,
+} = require('./schemas');
 
 const walletRegistrationPost = async function (req, res) {
   await walletRegistrationPostSchema.validateAsync(req.body, {
@@ -43,10 +21,8 @@ const walletRegistrationPost = async function (req, res) {
 
   const walletRegistrationService = new WalletRegistrationService();
 
-  const {
-    walletRegistration,
-    status,
-  } = await walletRegistrationService.createWalletRegistration(req.body);
+  const { walletRegistration, status } =
+    await walletRegistrationService.createWalletRegistration(req.body);
 
   res.status(status).json(walletRegistration);
 };
@@ -59,10 +35,11 @@ const walletRegistrationGet = async function (req, res) {
   const { filter, limitOptions } = getFilterAndLimitOptions(req.query);
 
   const walletRegistrationService = new WalletRegistrationService();
-  const walletRegistrations = await walletRegistrationService.getWalletRegistrations(
-    filter,
-    limitOptions,
-  );
+  const walletRegistrations =
+    await walletRegistrationService.getWalletRegistrations(
+      filter,
+      limitOptions,
+    );
   const count = await walletRegistrationService.getWalletRegistrationsCount(
     filter,
   );
@@ -89,9 +66,10 @@ const walletRegistrationSingleGet = async function (req, res) {
   });
 
   const walletRegistrationService = new WalletRegistrationService();
-  const walletRegistration = await walletRegistrationService.getWalletRegistrationById(
-    req.params.wallet_registration_id,
-  );
+  const walletRegistration =
+    await walletRegistrationService.getWalletRegistrationById(
+      req.params.wallet_registration_id,
+    );
 
   if (!walletRegistration?.id) {
     throw new HttpError(
