--- server/models/model.spec.js
@@ -1,1 +1,0 @@
-// empty spec
