+++ server/routes/index.js
@@ -1,0 +1,7 @@
+module.exports = [
+  require('./deviceConfigurationRoutes'),
+  require('./rawCaptureRoutes'),
+  require('./replayEventRoutes'),
+  require('./sessionRoutes'),
+  require('./walletRegistrationRoutes'),
+];
